+++ src/projections/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
 export * from './realtimeDBInlineProjection';
-export * from './realtimeDBInlineProjectionSpec';